ai.service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
ultralytics==8.0.206
Pillow==10.1.0
python-multipart==0.0.6
kafka-python==2.0.2
pyserial==3.5
apscheduler==3.10.4
